+++ src/utils/errors.ts
@@ -1,21 +1,41 @@
 import { ApiCallError } from 'iexec/errors';
 import { ValidationError } from 'yup';
+import { SafeObserver } from './reactive.js';
+
+export const updateErrorMessage = 'Failed to update oracle';
 
 export class WorkflowError extends Error {
-  constructor(message: string, cause: Error) {
+  isProtocolError: boolean;
+
+  constructor({
+    message,
+    cause,
+    isProtocolError = false,
+  }: {
+    message: string;
+    cause: Error;
+    isProtocolError?: boolean;
+  }) {
     super(message, { cause: cause });
     this.name = this.constructor.name;
+    this.isProtocolError = isProtocolError;
   }
 }
 
-export function handleProtocolError(error: Error): boolean {
+export function handleIfProtocolError(
+  error: Error,
+  observer: SafeObserver<unknown>
+) {
   if (error instanceof ApiCallError) {
-    throw new WorkflowError(
-      "A service in the iExec protocol appears to be unavailable. You can retry later or contact iExec's technical support for help.",
-      error
+    observer.error(
+      new WorkflowError({
+        message:
+          "A service in the iExec protocol appears to be unavailable. You can retry later or contact iExec's technical support for help.",
+        cause: error,
+        isProtocolError: true,
+      })
     );
   }
-  return false;
 }
 
 export class NoValueError extends Error {}